==> 0. news_crawling.ipynb <==
https://wonhwa.tistory.com/46

import pandas as pd
from bs4 import BeautifulSoup
import requests
import re
import time
from selenium import webdriver
from selenium.webdriver.common.by import By
from webdriver_manager.chrome import ChromeDriverManager
from selenium.common.exceptions import NoSuchElementException

##########검색어 리스트 생성################
negs_list = [ "손실", "위기", "부실", "하락", "감소", "리스크", "위험", "불확실성","삭감", "감축", "해고", 
             "제재", "벌금", "수사", "과징금", "실패", "파산", "배상", "손해", "검찰", "재판",
             "부실", "불법", "부적절", "위반", "적발", "횡령", "사기", "비리","논란","부도덕","처벌"]

negs_dict = {i: value for i, value in enumerate(negs_list, start=1)} # negs_dict 생성

def get_key(my_dict, val):
    for key, value in my_dict.items():
        if val == value:
            return key
    return "Value does not exist"



len(negs_list)


kospi = pd.read_excel("kis_코스피200.xlsx")

search_list = []

for i in range(len(kospi)):
    for j in range(len(negs_list)):
        search_list.append('"' + kospi.iloc[i] + '" ' + '+ ' + negs_list[j])

search_lst = [str(item[0]) for item in search_list] # search_lst 생성
search_lst


search_lst[31*147]

len(search_list)

kospi

oci_list = []

for j in range(len(negs_list)):
    oci_list.append('"' + 'OCI홀딩스' + '"' + ' + ' + negs_list[j])  # 따옴표와 공백 유지

oci_list


# 수정 코드

#크롤링시 필요한 라이브러리 불러오기
from bs4 import BeautifulSoup
import requests
from requests.adapters import HTTPAdapter
from requests.packages.urllib3.util.retry import Retry
import re
import datetime
from tqdm.auto import tqdm
import sys
import pandas as pd


session = requests.Session()
retry = Retry(total=5, backoff_factor=0.1, status_forcelist=[ 500, 502, 503, 504 ])
adapter = HTTPAdapter(max_retries=retry)
session.mount('http://', adapter)
session.mount('https://', adapter)

# 페이지 url 형식에 맞게 바꾸어 주는 함수 만들기
def makePgNum(num):
    if num == 1:
        return num
    elif num == 0:
        return num+1
    else:
        return num+9*(num-1)

def makeUrl(search, start_pg, end_pg):
    urls = []
    for i in range(start_pg, end_pg + 1):
        page = makePgNum(i)
        url = "https://search.naver.com/search.naver?where=news&sm=tab_pge&query=" + search + "&start=" + str(page)
        urls.append(url)
    return urls 

def news_attrs_crawler(articles,attrs):
    attrs_content=[]
    for i in articles:
        attrs_content.append(i.attrs[attrs])
    return attrs_content

# ConnectionError방지
headers = {"User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) Chrome/98.0.4758.102"}

def articles_crawler(url):
    original_html = requests.get(i,headers=headers)
    html = BeautifulSoup(original_html.text, "html.parser")
    url_naver = html.select("div.group_news > ul.list_news > li div.news_area > div.news_info > div.info_group > a.info")
    url = news_attrs_crawler(url_naver,'href')
    return url


# 시작 페이지와 종료 페이지
start_page = 1
end_page = 5

# 뉴스 크롤링 시작
news_titles = []
news_url =[]
news_contents =[]
news_dates = []
keywords = []

# 최종 데이터프레임 생성을 위한 빈 리스트 준비
keywords, news_titles, news_contents, news_dates = [], [], [], []
news_df = pd.DataFrame()

for idx, search in enumerate(tqdm(search_lst)):   # 키워드 리스트의 인덱스와 값을 동시에 사용
    urls = makeUrl(search, start_page, end_page)
    for i in urls:
        article_urls = articles_crawler(i)
        for url in article_urls:
            if "news.naver.com" in url:
                news = session.get(url, headers=headers, timeout=5)  # 타임아웃 설정
                news_html = BeautifulSoup(news.text, "html.parser")
                
                # 뉴스 제목 가져오기
                title = news_html.select_one("#ct > div.media_end_head.go_trans > div.media_end_head_title > h2")
                if title is None:
                    title = news_html.select_one("#content > div.end_ct > div > h2")

                # 뉴스 본문 가져오기
                content = news_html.select("div#dic_area")
                if content == []:
                    content = news_html.select("#articeBody")
                content = ''.join(str(content))

                # 기사 날짜 가져오기
                html_date = news_html.select_one("div#ct> div.media_end_head.go_trans > div.media_end_head_info.nv_notrans > div.media_end_head_info_datestamp > div > span")
                if html_date is not None:
                    news_date = html_date.attrs['data-date-time']
                else:
                    news_date = news_html.select_one("#content > div.end_ct > div > div.article_info > span > em")
                    news_date = re.sub(pattern='<[^>]*>', repl='', string=str(news_date))

                # html태그제거 및 텍스트 다듬기
                pattern1 = '<[^>]*>'
                title = re.sub(pattern=pattern1, repl='', string=str(title))
                content = re.sub(pattern=pattern1, repl='', string=content)
                pattern2 = """[\n\n\n\n\n// flash 오류를 우회하기 위한 함수 추가\nfunction _flash_removeCallback() {}"""
                content = content.replace(pattern2, '')

                keywords.append(search)
                news_titles.append(title)
                news_contents.append(content)
                news_dates.append(news_date)

                # 키워드별 데이터프레임 생성 및 저장
                news_df_temp = pd.DataFrame({'keyword': keywords, 'title': news_titles, 'content': news_contents, 'date': news_dates})
                news_df_temp = news_df_temp.drop_duplicates()

                # 합본
                news_df = pd.concat([news_df, news_df_temp])
                news_df = news_df.drop_duplicates()
                news_df.to_csv('temp_total_4.csv', index=False, encoding='utf-8-sig')
                
     # idx가 31의 배수가 될 때마다 데이터프레임을 저장합니다.
    if (idx + 1) % 31 == 0:
        # 파일명에 인덱스를 이용하여 저장합니다.
        news_df.to_csv('./temp/temp_df_{}.csv'.format((idx + 1) // 31), index=False, encoding='utf-8-sig')  # 31로 나눈 몫을 이용합니다.
    
    print(search + " done")
    print(len(news_df))

# 최종 결과 저장
#news_df = news_df.drop_duplicates()
#news_df.to_csv('final_news_df.csv', index=False, encoding='utf-8-sig')

# #데이터 프레임 저장
# now = datetime.datetime.now() 
# news_df.to_csv('{}_{}.csv'.format(search,now.strftime('%Y%m%d_%H시%M분%S초')),encoding='utf-8-sig',index=False)


news_df


#최종 결과 저장
news_df = news_df.drop_duplicates()
#news_df.to_csv('audit_data_final.csv', index=False, encoding='utf-8-sig')
