==> 1. text_preprocessing.ipynb <==
import pandas as pd

import os
os.getcwd()

df_raw = pd.read_csv('audit_data/audit_data_final.csv', index_col=0)
df_raw.head()

len(df_raw)

df = df_raw.drop_duplicates().dropna()

# 텍스트 전처리

#### 텍스트 전처리 ####
import re

def clean_text(text):
    # 리스트에 저장된 문자열 제거
    if text.startswith('[') and text.endswith(']'):
        text = text[1:-1]
    
    # HTML 태그 제거
    cleaned_text = re.sub('<.*?>', '', text)

    # HTML 코드 제거 (&amp; 등)
    cleaned_text = re.sub('&.*?;', '', cleaned_text)

    # 특수문자 및 개행 문자 제거
    cleaned_text = re.sub('[^가-힣0-9a-zA-Z\\s]', '', cleaned_text)
    cleaned_text = re.sub('\n', ' ', cleaned_text)
    cleaned_text = re.sub('\t', ' ', cleaned_text)
    
    return cleaned_text




# 텍스트 전처리 실행
df.loc[:,"cleaned_content"] = df.loc[:,'content'].apply(clean_text)
df.head()


df = df.drop(columns = 'content')
df.head()

df.to_csv('news_preprocessed_0620.csv', index=False)
