==> 2. model_kobert+klue.ipynb <==
import pandas as pd

import os
os.getcwd()

df = pd.read_csv('news_preprocessed_0620.csv')
df.head()

df = df.dropna()

# Kobert

import torch
import tensorflow as tf
from transformers import BertTokenizer, BertForSequenceClassification
from tqdm.auto import tqdm

print(f'Current device: {torch.cuda.get_device_name(torch.cuda.current_device())}')

if tf.config.list_physical_devices('GPU'):
    print("GPU is available")
else:
    print("GPU is not available")


# KOBERT

gpus = tf.config.experimental.list_physical_devices('GPU')
if gpus:
    try:
        # Currently, memory growth needs to be the same across GPUs
        for gpu in gpus:
            tf.config.experimental.set_memory_growth(gpu, True)
    except RuntimeError as e:
        # Memory growth must be set before GPUs have been initialized
        print(e)

# Check if a GPU is available and if not, use a CPU

device = torch.device("cuda")
        
# Load the pre-trained model
model = BertForSequenceClassification.from_pretrained("monologg/kobert").to(device)

# Load the pre-trained model tokenizer
tokenizer = BertTokenizer.from_pretrained("monologg/kobert")

# Create a new `output` column and initialize it to 0
df["output_kobert"] = 0.0

# Add a progress bar
pbar = tqdm(total=len(df))

for i, row in df.iterrows():
    # Input text
    text = row["cleaned_content"]

    # Tokenize the text with the tokenizer and convert it to the model's input format
    input_ids = tokenizer.encode(text, add_special_tokens=True)

    # Truncate the input_ids to 512 tokens if it's longer
    #if len(input_ids) > 512:
    #    input_ids = input_ids[:512]

    # Add padding or truncate to max length
    input_ids = input_ids[:512] if len(input_ids) > 512 else input_ids + [0] * (512 - len(input_ids))

    input_ids = torch.tensor([input_ids]).to(device)

    # Predict with the model
    outputs = model(input_ids)

    # Convert the result from log probabilities to probabilities
    probs = torch.nn.functional.softmax(outputs[0], dim=1)

    # Store the negative probability in the `output` column
    df.at[i, "output_kobert"] = probs[0][1].item()

    # Update the progress bar
    pbar.update()

# Close the progress bar
pbar.close()

# KLUE

device = torch.device("cuda")

# Load the pre-trained model and move it to the device
model = BertForSequenceClassification.from_pretrained("klue/bert-base").to(device) 

# Load the pre-trained model tokenizer
tokenizer = BertTokenizer.from_pretrained("klue/bert-base")

# Create a new `output` column and initialize it to 0
df["output_klue"] = 0.0

# Add a progress bar
pbar = tqdm(total=len(df))

for i, row in df.iterrows():
    # Input text
    text = row["cleaned_content"]

    # Tokenize the text with the tokenizer and convert it to the model's input format
    input_ids = tokenizer.encode(text, add_special_tokens=True)

    # Truncate the input_ids to 512 tokens if it's longer
    input_ids = input_ids[:512] if len(input_ids) > 512 else input_ids + [0] * (512 - len(input_ids))

    input_ids = torch.tensor([input_ids]).to(device)

    # Predict with the model
    outputs = model(input_ids)

    # Convert the result from log probabilities to probabilities
    probs = torch.nn.functional.softmax(outputs[0], dim=1)

    # Store the negative probability in the `output` column
    df.at[i, "output_klue"] = probs[0][1].item()

    # Update the progress bar
    pbar.update()

# Close the progress bar
pbar.close()

df.head()

# 추가 처리

# 문자열 분할
df['기업'], df['kwd'] = zip(*df['keyword'].apply(lambda x: x.split(' + ', 1)))

# 따옴표 제거
df['기업'] = df['기업'].str.strip('"')

df.head()

df.기업.unique()

df = df.rename(columns = {'output':'output_klue'})
df.head()

len(df)

df = df.dropna()
df.isnull().sum()

df = df.query('date != "None"')
df.head()

df['date_new'] = pd.to_datetime(df['date'], errors='coerce').dt.date # 날짜와 시간으로 변환
df.head()


df.to_csv('output_0620_v2.csv', index = False)
